Added fallbacks with optional write_through.
If redis returned nil for a particular lookup, it calls the
fallback proc.  If write_through is true, it writes the result
back in to redis.

If redis returns an error, such as a connection refused error,
it calls the fallback proc to get the result but ignores the
write_through option.